requirements.txt
Authlib==1.1.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.3
dnspython==2.2.1
email-validator==1.3.0
Flask==2.2.2
Flask-WTF==1.0.1
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.2
pycparser==2.21
pycryptodome==3.15.0
requests==2.28.1
urllib3==1.26.12
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.10.0
